Refactor cache funcs
